fix test failure

+++ cpp/src/arrow/util/async_generator.h
@@ -1499,7 +1499,7 @@ AsyncGenerator<T> MakeConcatenatedGenerator(AsyncGenerator<AsyncGenerator<T>> so
 template <typename T>
 struct Enumerated {
   T value;
-  int64_t index;
+  int index;
   bool last;
 };
